+++ lib.cjs/hds/hd.d.ts
@@ -3,7 +3,7 @@ import { HDOptionsInterface } from '../interfaces';
 import { EllipticCurveCryptography } from '../eccs';
 export declare class HD {
     ecc: typeof EllipticCurveCryptography;
-    protected derivation: any;
+    derivation: any;
     constructor(options?: HDOptionsInterface);
     static getName(): string;
     getName(): string;