+++ lib.cjs/hds/monero.js
@@ -68,6 +68,7 @@ class MoneroHD extends hd_1.HD {
         return this;
     }
     updateDerivation(derivation) {
+        this.cleanDerivation();
         return this.fromDerivation(derivation);
     }
     cleanDerivation() {