+++ lib.esm/hds/bip44.d.ts
@@ -10,8 +10,6 @@ export declare class BIP44HD extends BIP32HD {
     fromChange(change: string | number): this;
     fromAddress(address: number | [number, number]): this;
     fromDerivation(derivation: BIP44Derivation): this;
-    updateDerivation(derivation: BIP44Derivation): this;
-    cleanDerivation(): this;
     getAddress(options?: HDAddressOptionsInterface): string;
 }
 //# sourceMappingURL=bip44.d.ts.map