+++ lib.esm/hds/bip86.d.ts
@@ -5,8 +5,6 @@ export declare class BIP86HD extends BIP44HD {
     constructor(options?: HDOptionsInterface);
     static getName(): string;
     fromDerivation(derivation: BIP86Derivation): this;
-    updateDerivation(derivation: BIP86Derivation): this;
-    cleanDerivation(): this;
     getRootXPrivateKey(version?: Uint8Array | number, encoded?: boolean): string | null;
     getRootXPublicKey(version?: Uint8Array | number, encoded?: boolean): string | null;
     getXPrivateKey(version?: Uint8Array | number, encoded?: boolean): string | null;